LayoutTests/js/intl-invalid-locale-crash.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<meta charset="utf-8">
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
if (window.internals) {
    // Any language name with less than two characters is considered invalid, so we use "a" here.
    window.internals.setUserPreferredLanguages(["a"]);
}
shouldThrow("new Intl.DateTimeFormat().resolvedOptions()", "'TypeError: failed to initialize DateTimeFormat due to invalid locale'");
shouldThrow("new Intl.NumberFormat().resolvedOptions()", "'TypeError: failed to initialize NumberFormat due to invalid locale'");
shouldThrow("new Intl.Collator().resolvedOptions()", "'TypeError: failed to initialize Collator due to invalid locale'");
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
